2006-03-12  Russell M. Taylor II  <taylorr@cs.unc.edu>
	* time_test.cpp : Added type-cast.
	* vrpn_Keyboard_Mouse.C : Added required library.
	* vrpn_Tracker_DTrack.C : Changed from WIN to _WIN check.
	* vrpn.sln : Visual C++ 2005 solution file for VRPN.
	* *.vcproj : Visual C++ 2005 project files for VRPN.

+++ time_test.cpp
@@ -50,7 +50,7 @@ void	get_time_using_ftime(unsigned long &sec, unsigned long &usec)
 {
     struct _timeb t;
     _ftime(&t);
-    sec = t.time;
+    sec = static_cast<unsigned long>(t.time);
     usec = (long)t. millitm * 1000;
 }
